requirements.txt

streamlit>=1.36.0
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
SQLAlchemy>=2.0.30
apscheduler>=3.10.4
python-dotenv>=1.0.1
plotly>=5.22.0
requests>=2.32.3
